Move style encodings to be public
