+++ napari/layers/utils/style_encoding.py
